@@ -40,7 +40,7 @@ dependencies:
   provider: ^6.0.0
   geolocator: ^12.0.0
   geocoding: ^3.0.0
-  google_maps_flutter: ^2.6.1
+  google_maps_flutter: ^2.7.0
   qr_code_scanner: ^1.0.1
   firebase_core: ^2.31.0
   firebase_storage: ^11.7.5
@@ -50,8 +50,16 @@ dependencies:
   google_sign_in: ^6.2.1
   flutter_facebook_auth: ^7.0.0
   http: ^1.2.1
-  dio: ^4.0.0
+  dio: ^5.4.3+1
   location: ^6.0.2
+  url_launcher: ^6.3.0
+  webview_flutter: ^3.0.4
+  qr_flutter: ^4.1.0
+  flutter_barcode_scanner: ^2.0.0
+  barcode_widget: ^2.0.4
+  qr: ^3.0.1
+  image: ^3.0.1
+
 
 dev_dependencies:
   flutter_test:
@@ -601,6 +609,8 @@ flutter:
     - assets/vectors/incorrect_symbol_x2.svg
     - assets/vectors/back_16_x2.svg
     - assets/vectors/back_8_x2.svg
+    - assets/images/trainCar.png
+
     # An image asset can refer to one or more resolution-specific "variants", see
     # https://flutter.dev/assets-and-images/#resolution-aware
     # For details regarding adding assets from package dependencies, see
