@@ -3,7 +3,7 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:rail_ease/pages/forget.dart';
 
-import 'edit_profile .dart';
+import 'edit_profile.dart';
 
 void main() {
   runApp(MaterialApp(
